pyproject.toml
[tool.poetry]
name = "pluget"
version = "1.8.0"
description = "A powerful package manager which updates Plugins and Server Software for minecraft servers."
authors = ["Neocky", "An0nX"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
urllib3 = "^1.21.1"
requests = "^2.25.1"
paramiko = "<3.0.0"
bcrypt = "^3.2.0"
cryptography = "^3.4.6"
pynacl = "^1.4.0"
cffi = "^1.14.5"
six = "^1.15.0"
pycparser = "^2.20"
pysftp = "^0.2.9"
rich = "^9.13.0"
commonmark = "^0.9.1"
pygments = "^2.8.1"
typing-extensions = "^3.7.4.3"
ruamel-yaml = "^0.17.21"

[tool.poetry.group.dev.dependencies]
mypy = "^1.0.0"
black = "^23.0.0"
flake8 = "^6.0.0"
types-requests = "^2.28.0"
types-paramiko = "^3.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
